V8

+++ src/Controller/CategoryController.php
@@ -16,6 +16,10 @@ class CategoryController extends AbstractController
     #[Route('/', name: 'app_category_index', methods: ['GET'])]
     public function index(CategoryRepository $categoryRepository): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('home');
+        }
+        
         return $this->render('category/index.html.twig', [
             'categories' => $categoryRepository->findAll(),
         ]);
@@ -24,6 +28,10 @@ class CategoryController extends AbstractController
     #[Route('/new', name: 'app_category_new', methods: ['GET', 'POST'])]
     public function new(Request $request, CategoryRepository $categoryRepository): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('home');
+        }
+        
         $category = new Category();
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
@@ -43,6 +51,10 @@ class CategoryController extends AbstractController
     #[Route('/{id}', name: 'app_category_show', methods: ['GET'])]
     public function show(Category $category): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('home');
+        }
+        
         return $this->render('category/show.html.twig', [
             'category' => $category,
         ]);
@@ -51,6 +63,10 @@ class CategoryController extends AbstractController
     #[Route('/{id}/edit', name: 'app_category_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Category $category, CategoryRepository $categoryRepository): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('home');
+        }
+        
         $form = $this->createForm(CategoryType::class, $category);
         $form->handleRequest($request);
 
@@ -69,6 +85,10 @@ class CategoryController extends AbstractController
     #[Route('/{id}', name: 'app_category_delete', methods: ['POST'])]
     public function delete(Request $request, Category $category, CategoryRepository $categoryRepository): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('home');
+        }
+        
         if ($this->isCsrfTokenValid('delete'.$category->getId(), $request->request->get('_token'))) {
             $categoryRepository->remove($category, true);
         }
