+++ src/Controller/ResponseController.php
@@ -16,6 +16,10 @@ class ResponseController extends AbstractController
     #[Route('/', name: 'app_response_index', methods: ['GET'])]
     public function index(ResponseRepository $responseRepository): BaseResponse
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('home');
+        }
+        
         return $this->render('response/index.html.twig', [
             'responses' => $responseRepository->findAll(),
         ]);
@@ -24,6 +28,10 @@ class ResponseController extends AbstractController
     #[Route('/new', name: 'app_response_new', methods: ['GET', 'POST'])]
     public function new(Request $request, ResponseRepository $responseRepository): BaseResponse
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('home');
+        }
+        
         $response = new Response();
         $form = $this->createForm(ResponseType::class, $response);
         $form->handleRequest($request);
@@ -43,6 +51,10 @@ class ResponseController extends AbstractController
     #[Route('/{id}', name: 'app_response_show', methods: ['GET'])]
     public function show(Response $response): BaseResponse
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('home');
+        }
+        
         return $this->render('response/show.html.twig', [
             'response' => $response,
         ]);
@@ -51,6 +63,10 @@ class ResponseController extends AbstractController
     #[Route('/{id}/edit', name: 'app_response_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Response $response, ResponseRepository $responseRepository): BaseResponse
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('home');
+        }
+        
         $form = $this->createForm(ResponseType::class, $response);
         $form->handleRequest($request);
 
@@ -69,6 +85,10 @@ class ResponseController extends AbstractController
     #[Route('/{id}', name: 'app_response_delete', methods: ['POST'])]
     public function delete(Request $request, Response $response, ResponseRepository $responseRepository): BaseResponse
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('home');
+        }
+        
         if ($this->isCsrfTokenValid('delete'.$response->getId(), $request->request->get('_token'))) {
             $responseRepository->remove($response, true);
         }
